finishHomework06

+++ src/by/belhard/j20/ponkratov/homework/homework06/task04/PolygonCalc.java
@@ -2,8 +2,8 @@ package by.belhard.j20.ponkratov.homework.homework06.task04;
 
 public interface PolygonCalc {
 
-    int perimetr();
+    double perimetr();
 
-    int area();
+    double area();
 
 }
